Moved components to separate dir
