+++ src/game-loop/src/components/PhysicsComponent.cpp
@@ -1,4 +1,4 @@
-#include "PhysicsComponent.hpp"
+#include "components/PhysicsComponent.hpp"
 #include "MainDude.hpp"
 
 void PhysicsComponent::update(MainDude& main_dude)
